+++ readers/mutation_readers.cc
@@ -16,6 +16,7 @@
 #include "mutation/range_tombstone_splitter.hh"
 #include "readers/combined.hh"
 #include "readers/delegating_v2.hh"
+#include "readers/delegating_impl.hh"
 #include "readers/empty_v2.hh"
 #include "readers/mutation_reader.hh"
 #include "readers/forwardable_v2.hh"
