+++ test/boost/row_cache_test.cc
@@ -40,7 +40,7 @@
 
 #include <fmt/ranges.h>
 #include "readers/from_mutations_v2.hh"
-#include "readers/delegating_v2.hh"
+#include "readers/delegating_impl.hh"
 #include "readers/empty_v2.hh"
 #include <seastar/testing/thread_test_case.hh>
